♜♜♚♝
V/BWY29udmVydCBzdGF0ZSB0byBkaXJlY3RvcnkgbW9kdWxlLCBleHRyYWN0IHNjaGVtYSwgbGVnYWN5LCBhbmQgbG9hZGVyIChteCMxOTIgUFIgMTMvMTQp

[crc16:chess|snappy:base64]
